fix gcc warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@4721 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtlib/vtosg/SceneOSG.cpp
@@ -337,8 +337,8 @@ void vtScene::WorldToScreen(const FPoint3 &point, IPoint2 &result)
 	v2s(point, object);
 	Vec3 window;
 	m_pOsgSceneView->projectObjectIntoWindow(object, window);
-	result.x = window.x();
-	result.y = window.y();
+	result.x = (int) window.x();
+	result.y = (int) window.y();
 }
